+++ projects/components-demo/src/app/codeblock/codeblock.component.html
@@ -1,3 +1,4 @@
+<!-- TODO: Add a title bar for displaying the file's name above or within the codeblock. Only display if there is a name. -->
 <div class="code-wrapper">
   <ng-content></ng-content>
 </div>
