--- projects/components-demo/src/app/demo-multiselect/demo-multiselect.component.scss
@@ -1,5 +1,0 @@
-dp-button {
-  display: inline-flex;
-  margin-top: 0.5rem;
-  margin-right: 0.5rem;
-}
