+++ projects/components-demo/src/app/demos/demo-multiselect/demo-multiselect.component.html
@@ -1,0 +1,127 @@
+<div class="component-intro">
+  <h1>Multiselect Input</h1>
+  <p>
+    A selectable input component for choosing multiple values from a drop down menu. Utilizes the same component as the
+    <a routerLink="/select"><code>SelectComponent</code></a>. Supports <code>ControlValueAccessor</code>.
+  </p>
+</div>
+
+<div class="demo-wrapper">
+    <h4>Numbers</h4>
+    <dp-select placeholder="Please select" [formControl]="formControl" [options]="options" [multiselect]="true"></dp-select>
+    <div class="demo-errors" *ngIf="errors">{{ errors }}</div>
+
+    <div class="demo-btn-group">
+      <dp-button color="primary" *ngIf="this.formControl.disabled" (click)="enable()">Enable</dp-button>
+      <dp-button color="primary" *ngIf="!this.formControl.disabled" (click)="disable()">Disable</dp-button>
+      <dp-button color="secondary" (click)="error()">Error</dp-button>
+    </div>
+
+    <div class="demo-value">Value: {{ formControl.value }}</div>
+</div>
+
+<div class="component-info">
+  <dp-tab-group>
+    <dp-tab label="Documentation">
+      <h3>Getting Started</h3>
+      <p>To enable multi-select in the <code>SelectComponent</code>, add the <code>[multiselect]="true"</code> attribute.
+        The multi-select component can be used with Angular Form Control.</p>
+      <app-codeblock>{{ codeGetStarted }}</app-codeblock>
+      <h4>Import</h4>
+      <app-codeblock>Currently everything is imported by DevproComponentsModule.</app-codeblock>
+
+      <h3>Options</h3>
+      <p>The values contained the <code>SelectComponent</code>'s drop down can be easily passed in with the <code>options</code>
+      attribute. The options can be defined in a typescript file.</p>
+      <app-codeblock>{{ codeOptionsAttr }}</app-codeblock>
+      <app-codeblock>{{ codeOptions }}</app-codeblock>
+
+      <h3>Properties</h3>
+      <table>
+        <tr>
+          <th>Name</th>
+          <th>Type</th>
+          <th>Default Value</th>
+          <th>Description</th>
+        </tr>
+        <tr>
+          <td>placeholder</td>
+          <td>string</td>
+          <td>null</td>
+          <td>Placeholder for the input field.</td>
+        </tr>
+        <tr>
+          <td>multiselect</td>
+          <td>boolean</td>
+          <td>false</td>
+          <td>Must be set to true to enable multi-select.</td>
+        </tr>
+        <tr>
+          <td>error</td>
+          <td>boolean</td>
+          <td>false</td>
+          <td>Error state for error styling.</td>
+        </tr>
+        <tr>
+          <td>disabled</td>
+          <td>boolean</td>
+          <td>false</td>
+          <td>Disabled attribute of the select component, when true it will visibly change the input field and make it un-clickable.</td>
+        </tr>
+      </table>
+
+      <h3>Events</h3>
+      <table>
+        <tr>
+          <th>Name</th>
+          <th>Parameters</th>
+          <th>Description</th>
+        </tr>
+        <tr>
+          <td>notifyChanges</td>
+          <td>change: Emitted string that is a new change in selected value(s).</td>
+          <td>Notifies changes to the state of the field and its value(s).</td>
+        </tr>
+      </table>
+
+      <h3>Styling</h3>
+      <table>
+        <tr>
+          <th>Name</th>
+          <th>Element</th>
+        </tr>
+        <tr>
+          <td>select</td>
+          <td>Outer element containing the input box and its attributes.</td>
+        </tr>
+        <tr>
+          <td>selected-item-wrapper</td>
+          <td>Container element for the displaying the selected values.</td>
+        </tr>
+        <tr>
+          <td>filter</td>
+          <td>Styling for the filtering done on the options when searching.</td>
+        </tr>
+        <tr>
+          <td>item-wrapper</td>
+          <td>Outer element containing the items in the input field and the check beside already selected items.</td>
+        </tr>
+        <tr>
+          <td>item</td>
+          <td>Element displaying the options in the drop down.</td>
+        </tr>
+      </table>
+
+      <h3>Methods</h3>
+      <p>None.</p>
+
+      <h3>Dependencies</h3>
+      <p>None.</p>
+    </dp-tab>
+
+    <dp-tab label="Source Code">
+      <app-codeblock>{{ codeHTML }}</app-codeblock>
+      <app-codeblock>{{ codeTS }}</app-codeblock>
+    </dp-tab>
+  </dp-tab-group>
+</div>
