+++ projects/components-demo/src/app/demos/demo-multiselect/demo-multiselect.component.scss
@@ -1,0 +1,18 @@
+dp-button {
+  display: inline-flex;
+  margin-top: 0.5rem;
+  margin-right: 0.5rem;
+}
+
+dp-select {
+  width: 75%;
+}
+
+a {
+  color: rgba(59, 130, 246, 1);
+  text-decoration: none;
+}
+
+.demo-value {
+  width: 75%;
+}
